Improve: マスク切替部分の切り出し #136
